Restructure to clarify where fixture things are
